% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_version.R
\name{get_group_version}
\alias{get_group_version}
\title{Get the Library Version of a Zotero Group}
\usage{
get_group_version(group_id = 2352922, api_key = NULL)
}
\arguments{
\item{group_id}{The ID of the Zotero group.}

\item{api_key}{API key for Zotero. Only needed for private groups.}
}
\value{
The library version as an integer.
}
\description{
Returns the \code{Last-Modified-Version} of a Zotero group: an integer that increases
whenever anything in the group changes. Store it after a download and compare it
with the current value to decide whether a re-download is necessary.
}
\examples{
\dontrun{
get_group_version(2352922)
}

}
