% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_from_name.R
\name{id_from_name}
\alias{id_from_name}
\title{Get Zotero User ID from Username}
\usage{
id_from_name(username)
}
\arguments{
\item{username}{The Zotero username. \strong{Zotero Usernames are case sensitive!}}
}
\value{
The Zotero user ID as a string.
}
\description{
This function retrieves a Zotero user ID from a given username.
}
\details{
\strong{NB:The function relies on analysing the profile page. As this page might change,
it is not recommended to use this function programmatically, but rather to retrieve
the id and to use it hardcoded.}
}
\examples{
\dontrun{
id_from_name("ipbes")
# "5760254"
}
}
