% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groupids_from_user.R
\name{get_groupids_from_user}
\alias{get_groupids_from_user}
\title{Get all the ids from all Groups from a User}
\usage{
get_groupids_from_user(
  zotero_user_name = "ipbes",
  zotero_user_id = NULL,
  api_key = NULL,
  verbose = FALSE
)
}
\arguments{
\item{zotero_user_name}{Name of the user to download the groups from.
If zotero_user_id is specified, not needed. Default: "ipbes".}

\item{zotero_user_id}{Zotero user id to download the groups from. Default: obtained
through the function [\link{id_from_name}]{@link id_from_name}.
NB: In programatically usage, the id f=should be specified sxplicitely. Please see the
documentation for \link{id_from_name} for details.}

\item{api_key}{Zotero API key - only needed for private groups. Only read access needed.}
}
\value{
Named vector with the group ids, and the names of the groups.
}
\description{
#' Get the group ids of the Groups of the user with the zotero_user_name
This function requires a Zotero API key.
After logging into Zotero, go to the following page:
}
\details{
\href{https://www.zotero.org/settings/security#applications}{Home > Settings > Secutrity}

and \strong{Create New Private Key} with the following permissions:

Personal Library
- Allow library access
- Allow group access

Default Group Permissions
- Read Only
}
