% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups_from_user.R
\name{get_groups_from_user}
\alias{get_groups_from_user}
\title{Get all Groups from a User}
\usage{
get_groups_from_user(
  zotero_user_name = "ipbes",
  zotero_user_id = NULL,
  path = tempfile(),
  output_format = "zotero_rdf",
  api_key = Sys.getenv("ZOTERO_API_IPBES")
)
}
\arguments{
\item{zotero_user_name}{Name of the user to download the groups from. Default: "ipbes".}

\item{zotero_user_id}{Zotero user id to download the groups from. Default: obtained
through the function [\link{id_from_name}]{@link id_from_name}.
NB: In programatically usage, the id f=should be specified sxplicitely. Please see the
documentation for \link{id_from_name} for details.}

\item{output_format}{The output_format of the files. See \code{\link{zotero_get_group}}}

\item{api_key}{Zotero API key - only needed for private groups. Only read access needed.}
}
\description{
Download of the Groups of the user with the zotero_user_id.
This function requires a Zotero API key.
After logging into Zotero, go to the following page:
}
\details{
\href{https://www.zotero.org/settings/security#applications}{Home > Settings > Secutrity}

and \strong{Create New Private Key} with the following permissions:

Personal Library
- Allow library access
- Allow group access

Default Group Permissions
- Read Only
}
